refactor: rename storefront component directories to 'storefront' instead of 'store'
This is to avoid confusion with the redux store directory
